Service e erro de no such element

+++ organize/src/main/java/br/com/bamt/organize/controller/CompraController.java
@@ -3,16 +3,15 @@ package br.com.bamt.organize.controller;
 import br.com.bamt.organize.controller.dto.CompraDto;
 import br.com.bamt.organize.controller.form.NovaContaForm;
 import br.com.bamt.organize.model.Compra;
-import br.com.bamt.organize.model.Estabelecimento;
 import br.com.bamt.organize.model.repository.CompraRepository;
+import br.com.bamt.organize.service.CompraService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.time.LocalDate;
-import java.util.Locale;
 import java.util.Optional;
 
 @RestController
@@ -21,29 +20,30 @@ public class CompraController {
 
     @Autowired
     private CompraRepository compraRepository;
+    @Autowired
+    private CompraService compraService;
 
     @PostMapping("nova")
     @Transactional
-    public CompraDto novaCompra(@RequestBody @Valid NovaContaForm novaContaForm){
-        Compra compra = novaContaForm.toCompra();
-        compraRepository.save(compra);
-        return new CompraDto(compra);
+    public CompraDto salvarCompra(@RequestBody @Valid NovaContaForm novaContaForm){
+        return compraService.salvaCompra(novaContaForm);
     }
 
     @PostMapping("edita/{id}")
     @Transactional
-    public CompraDto editaCompra(@RequestBody @Valid NovaContaForm novaContaForm, @PathVariable Long id){
-        CompraDto compra = novaContaForm.atualizar(id, compraRepository);
-        return compra;
+    public CompraDto editarCompra(@RequestBody @Valid NovaContaForm novaContaForm, @PathVariable Long id){
+        return compraService.editaCompra(novaContaForm, id);
     }
 
     @PostMapping("deleta/{id}")
     @Transactional
-    public String deletaCompra(@PathVariable Long id){
+    public ResponseEntity<?> deletaCompra(@PathVariable Long id){
         Optional<Compra> compra = compraRepository.findById(id);
-        String nomeDaCompra = compra.get().getNomeEstabelecimento();
-        compraRepository.delete(compra.get());
-        return "A compra" + nomeDaCompra + "foi excluída com sucesso";
+        if(compra.isPresent()){
+            compraRepository.delete(compra.get());
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
